api/requirements.txt
fastapi
uvicorn
python-multipart
pillow
matplotlib
numpy
tensorflow-serving-api==2.5.0

